Change type of networkTime: uint32_t -> uint64_t.

+++ MAC/include/ControlPacket.h
@@ -23,20 +23,20 @@
 
 typedef struct 
 {
-    uint16_t     nodeID;
+    uint16_t    nodeID;
     uint32_t    currentFrame;
     uint8_t     currentSlot;
     uint8_t     collisionSlot;
     uint32_t    collisionFrequency;
     uint8_t     hopCount;
-    uint32_t    networkTime;
+    uint64_t    networkTime;
     uint32_t    initTime;
     uint8_t     ack;
 } ControlPacket_t;
 
 void controlpacket_init(ControlPacket_t DOUBLE_POINTER pkt);
 void controlpacket_destroy(ControlPacket_t DOUBLE_POINTER pkt);
-void controlpacket_create(ControlPacket_t *pkt, uint16_t nodeID, uint32_t frame, uint8_t slot, uint8_t collisionSlots, uint32_t collisionFrequency, uint8_t hopCount, uint32_t netTime, uint32_t initTime, uint8_t ack);
+void controlpacket_create(ControlPacket_t *pkt, uint16_t nodeID, uint32_t frame, uint8_t slot, uint8_t collisionSlots, uint32_t collisionFrequency, uint8_t hopCount, uint64_t netTime, uint32_t initTime, uint8_t ack);
 void controlpacket_clear(ControlPacket_t *pkt);
 void controlpacket_set_nodeID(ControlPacket_t *pkt, uint16_t nodeID);
 uint16_t controlpacket_get_nodeID(ControlPacket_t *pkt);
@@ -50,8 +50,8 @@ void controlpacket_set_collision_frequency(ControlPacket_t *pkt, uint32_t freque
 uint32_t controlpacket_get_collision_frequency(ControlPacket_t *pkt);
 void controlpacket_set_hop_count(ControlPacket_t *pkt, uint8_t count);
 uint8_t controlpacket_get_hop_count(ControlPacket_t *pkt);
-void controlpacket_set_network_time(ControlPacket_t *pkt, uint32_t time);
-uint32_t controlpacket_get_network_time(ControlPacket_t *pkt);
+void controlpacket_set_network_time(ControlPacket_t *pkt, uint64_t time);
+uint64_t controlpacket_get_network_time(ControlPacket_t *pkt);
 void controlpacket_set_init_time(ControlPacket_t *pkt, uint32_t time);
 uint32_t controlpacket_get_init_time(ControlPacket_t *pkt);
 void controlpacket_set_ACK(ControlPacket_t *pkt, uint8_t ack);
